Implement Product mutations and Queries

+++ src/main/java/com/alpha/orderservice/dto/ProductDto.java
@@ -7,6 +7,6 @@ import java.math.BigDecimal;
 public class ProductDto {
     private Long id;
     private String name;
-    private Long stock;
-    private BigDecimal price;
+    private int stock;
+    private double price;
 }
