+++ src/main/java/com/alpha/orderservice/entity/Product.java
@@ -19,6 +19,6 @@ public class Product {
     @GeneratedValue(strategy = GenerationType.AUTO)
     private Long id;
     private String name;
-    private Long stock;
-    private BigDecimal price;
+    private int stock;
+    private double price;
 }
